Validate Jot fields
This commit adds a line of code to the jot model which validates that
the jot name or content cannot be left blank. If blank, a 422 is
returned and the jot is not created, as intended.

Note: In the future, there really should be a way to have a blank jot
when jots are infact only lines within a thought. that way thoughts can
be organized with spacers as needed. Perhaps these spacing elements
should not be considered jots - it is useless to save a bunch of
spacers in the api.